ADD & MOD: manage favo num in each comment
